Create a video driver for EVKv2-lite zynq project, with CCAM5 Huahine
The driver mostly manages allocations, transfers using DMA engine API, and
answers mandatory V4L2 ioctls. Besides the DMA (using stock Xilinx IP), the
control of both Prophesee soft IP in the FPGA, and Huahine on CCAM5 module
are done with a copy of register access sequences extracts from sensorlib.
Values are never checked and a delay is systematically added between each
access to allow hardware to manage it (for PLL lock time, for instance).
In this state, the driver can't manage a full EVKv2 implementation, and can
only manage a CCAM5 Huahine (but doesn't check if the sensor module is one)

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>